+++ utils/deck/esbuild.js
@@ -1,9 +1,19 @@
 const esbuild = require('esbuild');
+const {readdirSync, existsSync, mkdirSync, writeFileSync} = require('fs');
+const {join} = require('path');
+
+const lib = join(process.cwd(), 'lib');
+
+if (!existsSync(lib)) {
+  mkdirSync(lib);
+}
+
+const entryPoints = readdirSync(join(process.cwd(), 'src')).map((file) => `src/${file}`);
 
 esbuild
   .build({
-    entryPoints: ['src/index.ts'],
-    outdir: 'lib',
+    entryPoints,
+    outdir: 'lib/esm',
     bundle: true,
     sourcemap: true,
     minify: true,
@@ -12,3 +22,19 @@ esbuild
     target: ['esnext']
   })
   .catch(() => process.exit(1));
+
+esbuild
+  .build({
+    entryPoints: ['src/index.ts'],
+    outfile: 'lib/cjs/index.cjs.js',
+    bundle: true,
+    sourcemap: true,
+    minify: true,
+    platform: 'node',
+    target: ['node16']
+  })
+  .catch(() => process.exit(1));
+
+writeFileSync(join(lib, 'index.js'), "export * from './esm/index.js';");
+
+writeFileSync(join(lib, 'index.cjs.js'), "module.exports = require('./cjs/index.cjs.js');");