+++ webcomponents/core/stencil.config.ts
@@ -16,7 +16,7 @@ export const config: Config = {
       serviceWorker: null
     },
     {
-      type: 'dist-custom-elements-bundle'
+      type: 'dist-custom-elements'
     }
   ],
   plugins: [