+++ CHANGELOG.md
@@ -4,7 +4,12 @@
 
 ### Web Components:
 
+- core: v10.0.0 ([CHANGELOG](https://github.com/deckgo/deckdeckgo/blob/main/webcomponents/core/CHANGELOG.md))
 - monaco-editor: v1.1.0 ([CHANGELOG](https://github.com/deckgo/deckdeckgo/blob/main/webcomponents/monaco-editor/CHANGELOG.md))
+
+### Others
+
+- deck-utils: v7.0.0 ([CHANGELOG](https://github.com/deckgo/deckdeckgo/blob/main/utils/deck/CHANGELOG.md))
 
 <a name="6.0.0"></a>